+++ pkgs/google_generative_ai/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - Add support for parsing Vertex AI specific fields in `CountTokensResponse`.
 - Add named constructors on `Schema` for each value type.
+- Add `GenerationConfig.responseMimeType` which supports setting
+  `'application/json'` to force the model to reply with JSON parseable output.
+- Add `outputDimensionality` argument support for `embedContent` and
+  `batchEmbedContent`.
 
 ## 0.3.3